+++ matdb/atoms.py
@@ -12,13 +12,11 @@ from itertools import product
 import h5py
 from ase import io
 from six import string_types
-import lazy_import
 from os import path
 from uuid import uuid4
 from matdb import msg
 from matdb.transforms import conform_supercell
-
-calculators = lazy_import.lazy_module("matdb.calculators")
+import matdb.calculators as calculators
 
 def _recursively_convert_units(in_dict, split = False):
     """Recursively goes through a dictionary and converts it's units to be
@@ -390,7 +388,12 @@ class Atoms(ase.Atoms):
             atoms (matdb.atoms.Atoms): the atoms object to perform calculations
             on.
         """
-        import quippy
+        try:
+            import quippy
+        except:
+            msg.err("Could not load quippy.")
+            return None
+
         props = self.properties.copy()
         params = self.params.copy()
         for k, v in props.items():