+++ matdb/utility.py
@@ -33,7 +33,6 @@ import h5py
 from contextlib import contextmanager
 
 from matdb import __version__
-from matdb.atoms import AtomsList
 # from matdb.database.utility import dbconfig
 # from matdb.utility import special_functions
 
@@ -181,7 +180,7 @@ def execute(args, folder, wait=True, nlines=100, venv=None,
             if len(error) >= nlines:
                 break
         pexec.stderr.close()
-        if printerr and len(error) > 0:
+        if printerr and len(error) > 0 and all([isinstance(i, string_types) for i in error]):
             msg.err(''.join(error))
 
     return {
@@ -198,6 +197,8 @@ def h5cat(files, target):
         target (str): name/path of the output file that will include all of the
           combined files.
     """
+    # Local import to prevent cyclic imports
+    from matdb.atoms import AtomsList
     result = AtomsList()
     for fname in files:
         ilist = AtomsList(fname)
@@ -451,10 +452,6 @@ def pgrid(options, ignore=None):
 
     grid = list(product(*values))
     return (grid, keys)
-
-import pytz
-from datetime import datetime
-from six import string_types
 
 epoch = datetime(1970,1,1, tzinfo=pytz.utc)
 """datetime.datetime: 1/1/1970 for encoding UNIX timestamps.
@@ -1029,3 +1026,64 @@ def required_packages():
             "numpy", "phenum", "phonopy", "pyparsing", "python-dateutil", "pytz",
             "PyYAML", "requests", "six", "subprocess32", "termcolor",
             "tqdm", "urllib3", "webencodings", "lazy-import", "seekpath"]
+
+def recursive_getattr(o, fqn):
+    """Recursively gets a '.'-separated path of attributes from an object.
+    
+    Args:
+        fqn (str): '.'-separated attribute path.
+    """
+    if '.' in fqn:
+        attrs = fqn.split('.')
+        for cattr in attrs[0:-1]:
+            if isinstance(o, dict) and cattr in o:
+                o = o[cattr]
+            elif hasattr(o, cattr):
+                o = getattr(o, cattr)
+            else:
+                return None
+    else:
+        attrs = [fqn]
+    
+    if isinstance(o, dict) and attrs[-1] in o:
+        return o[attrs[-1]]
+    elif hasattr(o, attrs[-1]):
+        return getattr(o, attrs[-1])        
+    else:
+        return None
+
+def recursive_setattr(o, fqn, value):
+    """Recursively sets a '.'-separated path of attributes from an object.
+    
+    Args:
+        fqn (str): '.'-separated attribute path.
+    """
+    if '.' in fqn:
+        attrs = fqn.split('.')
+        for cattr in attrs[0:-1]:
+            if isinstance(o, dict) and cattr in o:
+                o = o[cattr]
+            elif hasattr(o, cattr):
+                o = getattr(o, cattr)
+            else:
+                return None
+    else:
+        attrs = [fqn]
+    
+    if isinstance(o, dict):
+        o[attrs[-1]] = value
+    else:
+        setattr(o, attrs[-1], value)
+
+config_specs = {}
+def _set_config_paths(name, cntr_dir):
+    """Creates a dictionary of the matdb contrallor name and controll directory.
+
+    Args:
+        name (str): 'title' from the YML file.
+        cntr_dir (str): the path to the control directory.    
+    """
+
+    global config_specs  
+    config_specs["name"] = name
+    config_specs["cntr_dir"] = cntr_dir